Cleaned up build process
The old makefile/build was an absolute trainwreck, so I redid the
whole thing. The build is relatively sane now.

Additionally, generated files have been renamed from their defaults
to sensible names which match the project.

Finally, the main method was moved into the parser from it's own
file. I will probably split it back out later, but for the moment
the focus in on the scanner so I don't really care.