Implemented basic scanner
A subset of the c- scanner is completed. There are still tokens
which are not properly identified and a few tokens which are
identified in the wrong places, but the general layout of the
scanner is complete, as well as the passing of token structures
between the parser and scanner.

Known issues:
  CHARCONST tokens are not recognized at all.
  NUMCONST tokens have the wrong value in the token.value attribute.