Implemented all tokens
Implemented recognition and printing of all token types in the given
test file. I still need to go through the c- grammar and make sure
that the list of tokens I've implemented is the actual complete list
of tokens.

Furthermore, the version of the test file currently checked in is out
of date. There's at a minimum a formatting error and "in" is not a
keyword in the F16 version of c-, even though it is marked as such
in the test output. I'm having internet connectivity issues at the
moment so I can't pull down the updated test file, but that should
be updated once I have a stable connection.